requirements.txt
# CyberDyne — Dependências
# Instale com: pip install -r requirements.txt

# Requisições HTTP e parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Terminal colorido
colorama>=0.4.6

# Geração de PDF
reportlab>=4.0.0

# Análise de tokens JWT
PyJWT>=2.8.0

# Consultas DNS (subdomain takeover)
dnspython>=2.4.0

# Criptografia e verificação de certificados TLS
cryptography>=42.0.0

# Verificação de versões de pacotes
packaging>=24.0

# Variáveis de ambiente
python-dotenv>=1.0.0
